requirements.txt
pandas
Flask
flasgger
numpy
python-dotenv
strawberry-graphql[flask]
